Move NoteDialog to notes
